RF_11.003.04 | ManageJenkinsPageTest> testNavigateToSystemConfigurationPagesByEnter
